+++ op-node/rollup/driver/step.go
@@ -9,8 +9,6 @@ import (
 	"github.com/ethereum-optimism/optimism/op-node/rollup"
 	"github.com/ethereum-optimism/optimism/op-node/rollup/derive"
 
-	"github.com/ethereum/go-ethereum/common/hexutil"
-	"github.com/ethereum/go-ethereum/core/types"
 	"github.com/ethereum/go-ethereum/log"
 )
 
@@ -27,70 +25,16 @@ func (d *outputImpl) createNewBlock(ctx context.Context, l2Head eth.L2BlockRef, 
 	fetchCtx, cancel := context.WithTimeout(ctx, time.Second*20)
 	defer cancel()
 
-	var l1Info eth.L1Info
-	var receipts types.Receipts
-	var err error
-
-	seqNumber := l2Head.SequenceNumber + 1
-
-	// If the L1 origin changed this block, then we are in the first block of the epoch. In this
-	// case we need to fetch all transaction receipts from the L1 origin block so we can scan for
-	// user deposits.
-	if l2Head.L1Origin.Number != l1Origin.Number {
-		if l2Head.L1Origin.Hash != l1Origin.ParentHash {
-			d.log.Error("SEQUENCING BUG: cannot create new block on top of l2Head with new origin", "head", l2Head,
-				"head_origin", l2Head.L1Origin, "head_seq_nr", l2Head.SequenceNumber, "new_origin", l1Origin, "new_origin_parent", l1Origin.ParentID())
-			return l2Head, nil, fmt.Errorf("cannot create new block with L1 origin %s (parent %s) on top of L1 origin %s", l1Origin, l1Origin.ParentID(), l2Head.L1Origin)
-		}
-		l1Info, _, receipts, err = d.dl.Fetch(fetchCtx, l1Origin.Hash)
-		seqNumber = 0 // reset sequence number at the start of the epoch
-	} else {
-		if l2Head.L1Origin.Hash != l1Origin.Hash {
-			d.log.Error("SEQUENCING BUG: cannot create new block on top of l2Head with different origin at same height",
-				"head", l2Head, "head_origin", l2Head.L1Origin, "head_seq_nr", l2Head.SequenceNumber, "new_origin", l1Origin, "new_origin_parent", l1Origin.ParentID())
-			return l2Head, nil, fmt.Errorf("cannot create new block with L1 origin %s (parent %s) on top of L1 origin %s", l1Origin, l1Origin.ParentID(), l2Head.L1Origin)
-		}
-		l1Info, err = d.dl.InfoByHash(fetchCtx, l1Origin.Hash)
-	}
-	if err != nil {
-		return l2Head, nil, fmt.Errorf("failed to fetch L1 block info of %s: %v", l1Origin, err)
-	}
-
-	// Start building the list of transactions to include in the new block.
-	var txns []eth.Data
-
-	// First transaction in every block is always the L1 info transaction.
-	l1InfoTx, err := derive.L1InfoDepositBytes(seqNumber, l1Info)
+	attrs, _, err := derive.PreparePayloadAttributes(fetchCtx, d.Config, d.dl, l2Head, l1Origin.ID())
 	if err != nil {
 		return l2Head, nil, err
 	}
-	txns = append(txns, l1InfoTx)
-
-	// Next we append user deposits. If we're not the first block in an epoch, then receipts will
-	// be empty and no deposits will be derived.
-	deposits, errs := derive.DeriveDeposits(receipts, d.Config.DepositContractAddress)
-	d.log.Info("Derived deposits", "deposits", deposits, "l2Parent", l2Head, "l1Origin", l1Origin)
-	for _, err := range errs {
-		d.log.Error("Failed to derive a deposit", "l1OriginHash", l1Origin.Hash, "err", err)
-	}
-	// TODO: Should we halt if len(errs) > 0? Opens up a denial of service attack, but prevents lockup of funds.
-	txns = append(txns, deposits...)
 
 	// If our next L2 block timestamp is beyond the Sequencer drift threshold, then we must produce
 	// empty blocks (other than the L1 info deposit and any user deposits). We handle this by
 	// setting NoTxPool to true, which will cause the Sequencer to not include any transactions
 	// from the transaction pool.
-	nextL2Time := l2Head.Time + d.Config.BlockTime
-	shouldProduceEmptyBlock := nextL2Time >= l1Origin.Time+d.Config.MaxSequencerDrift
-
-	// Put together our payload attributes.
-	attrs := &eth.PayloadAttributes{
-		Timestamp:             hexutil.Uint64(nextL2Time),
-		PrevRandao:            eth.Bytes32(l1Info.MixDigest()),
-		SuggestedFeeRecipient: d.Config.FeeRecipientAddress,
-		Transactions:          txns,
-		NoTxPool:              shouldProduceEmptyBlock,
-	}
+	attrs.NoTxPool = uint64(attrs.Timestamp) >= l1Origin.Time+d.Config.MaxSequencerDrift
 
 	// And construct our fork choice state. This is our current fork choice state and will be
 	// updated as a result of executing the block based on the attributes described above.