+++ packages/contracts-bedrock/contracts/libraries/rlp/RLPWriter.sol
@@ -2,17 +2,18 @@
 pragma solidity ^0.8.9;
 
 /**
+ * @custom:attribution https://github.com/bakaoh/solidity-rlp-encode
  * @title RLPWriter
- * @author Bakaoh (with modifications)
+ * @author RLPWriter is a library for encoding Solidity types to RLP bytes. Adapted from Bakaoh's
+ *         RLPEncode library (https://github.com/bakaoh/solidity-rlp-encode) with minor
+ *         modifications to improve legibility.
  */
 library RLPWriter {
-    /**********************
-     * Internal Functions *
-     **********************/
-
     /**
-     * RLP encodes a byte string.
+     * @notice RLP encodes a byte string.
+     *
      * @param _in The byte string to encode.
+     *
      * @return The RLP encoded string in bytes.
      */
     function writeBytes(bytes memory _in) internal pure returns (bytes memory) {
@@ -28,8 +29,10 @@ library RLPWriter {
     }
 
     /**
-     * RLP encodes a list of RLP encoded byte byte strings.
+     * @notice RLP encodes a list of RLP encoded byte byte strings.
+     *
      * @param _in The list of RLP encoded byte strings.
+     *
      * @return The RLP encoded list of items in bytes.
      */
     function writeList(bytes[] memory _in) internal pure returns (bytes memory) {
@@ -38,8 +41,10 @@ library RLPWriter {
     }
 
     /**
-     * RLP encodes a string.
+     * @notice RLP encodes a string.
+     *
      * @param _in The string to encode.
+     *
      * @return The RLP encoded string in bytes.
      */
     function writeString(string memory _in) internal pure returns (bytes memory) {
@@ -47,8 +52,10 @@ library RLPWriter {
     }
 
     /**
-     * RLP encodes an address.
+     * @notice RLP encodes an address.
+     *
      * @param _in The address to encode.
+     *
      * @return The RLP encoded address in bytes.
      */
     function writeAddress(address _in) internal pure returns (bytes memory) {
@@ -56,8 +63,10 @@ library RLPWriter {
     }
 
     /**
-     * RLP encodes a uint.
+     * @notice RLP encodes a uint.
+     *
      * @param _in The uint256 to encode.
+     *
      * @return The RLP encoded uint256 in bytes.
      */
     function writeUint(uint256 _in) internal pure returns (bytes memory) {
@@ -65,8 +74,10 @@ library RLPWriter {
     }
 
     /**
-     * RLP encodes a bool.
+     * @notice RLP encodes a bool.
+     *
      * @param _in The bool to encode.
+     *
      * @return The RLP encoded bool in bytes.
      */
     function writeBool(bool _in) internal pure returns (bytes memory) {
@@ -75,14 +86,12 @@ library RLPWriter {
         return encoded;
     }
 
-    /*********************
-     * Private Functions *
-     *********************/
-
     /**
-     * Encode the first byte, followed by the `len` in binary form if `length` is more than 55.
-     * @param _len The length of the string or the payload.
+     * @notice Encode the first byte and then the `len` in binary form if `length` is more than 55.
+     *
+     * @param _len    The length of the string or the payload.
      * @param _offset 128 if item is string, 192 if item is list.
+     *
      * @return RLP encoded bytes.
      */
     function _writeLength(uint256 _len, uint256 _offset) private pure returns (bytes memory) {
@@ -110,9 +119,10 @@ library RLPWriter {
     }
 
     /**
-     * Encode integer in big endian binary form with no leading zeroes.
-     * @notice TODO: This should be optimized with assembly to save gas costs.
+     * @notice Encode integer in big endian binary form with no leading zeroes.
+     *
      * @param _x The integer to encode.
+     *
      * @return RLP encoded bytes.
      */
     function _toBinary(uint256 _x) private pure returns (bytes memory) {
@@ -134,11 +144,12 @@ library RLPWriter {
     }
 
     /**
-     * Copies a piece of memory to another location.
-     * @notice From: https://github.com/Arachnid/solidity-stringutils/blob/master/src/strings.sol.
+     * @custom:attribution https://github.com/Arachnid/solidity-stringutils
+     * @notice Copies a piece of memory to another location.
+     *
      * @param _dest Destination location.
-     * @param _src Source location.
-     * @param _len Length of memory to copy.
+     * @param _src  Source location.
+     * @param _len  Length of memory to copy.
      */
     function _memcpy(
         uint256 _dest,
@@ -169,9 +180,11 @@ library RLPWriter {
     }
 
     /**
-     * Flattens a list of byte strings into one byte string.
-     * @notice From: https://github.com/sammayo/solidity-rlp-encoder/blob/master/RLPEncode.sol.
+     * @custom:attribution https://github.com/sammayo/solidity-rlp-encoder
+     * @notice Flattens a list of byte strings into one byte string.
+     *
      * @param _list List of byte strings to flatten.
+     *
      * @return The flattened byte string.
      */
     function _flatten(bytes[] memory _list) private pure returns (bytes memory) {