+++ packages/contracts-bedrock/contracts/test/ResourceMetering.t.sol
@@ -7,6 +7,10 @@ import { Proxy } from "../universal/Proxy.sol";
 
 contract MeterUser is ResourceMetering {
     constructor() {
+        initialize();
+    }
+
+    function initialize() public initializer {
         __ResourceMetering_init();
     }